Safari-609.1.11.1/LayoutTests/editing/pasteboard/entries-api/DirectoryEntry-getFile-back-forward-cache-expected.txt

Make sure that fileSystemDirectoryEntry.getFile() does not prevent a page from entering the back/forward cache.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


pageshow - not from cache
pagehide - entering cache
pageshow - from cache
PASS Page did enter and was restored from the page cache
PASS Success callback was called
PASS restoredFromPageCache is true
PASS successfullyParsed is true

TEST COMPLETE
